Updated the example in the README file to be a little cleaner, and to include a TraceListener.
Moved the existing fake unit tests to an example source tree. This serves as an example and a visual set of test to check the output graphics look OK.

Added an actual unit test to make sure wrapping works as expected.

Moved the tooltip label into a new component class which also allows for multi-line comments.

Status text is shown at the top, and is coloured according to whether or not the test run was successful.

Updated the SWC so that it only includes the necessary classes.